Moves files from Prac07 into Prac08 directory; Removes unnecessary class statements from start of files, adds imports as replacement. Adds files car and usedCars to Prac07
