--- Prac07/unreliableCar.py
@@ -1,56 +1,0 @@
-from random import randint
-
-class Car:
-    """ represent a car object """
-
-    def __init__(self, name="", fuel=0):
-        """ initialise a Car instance """
-        self.name = name
-        self.fuel = fuel
-        self.odometer = 0
-
-    def __str__(self):
-        return "{}, fuel={}, odo={}".format(self.name, self.fuel, self.odometer)
-
-    def add_fuel(self, amount):
-        """ add amount to the car's fuel"""
-        self.fuel += amount
-
-    def drive(self, distance):
-        """ drive the car a given distance if it has enough fuel or drive until fuel runs out
-        return the distance actually driven """
-        if distance > self.fuel:
-            distance_driven = self.fuel
-            self.fuel = 0
-        else:
-            self.fuel -= distance
-            distance_driven = distance
-        self.odometer += distance_driven
-        return distance_driven
-
-
-class UnreliableCar(Car):
-    """docstring lol"""
-    def __init__(self, name, fuel, reliability):
-        super().__init__(name, fuel)
-        self.reliability = reliability
-
-    def drive(self, distance):
-        reliabilityTest = randint(0, 100)
-        print(reliabilityTest)
-        if self.reliability < reliabilityTest:
-            distance_driven = 0
-        elif distance > self.fuel:
-            distance_driven = self.fuel
-            self.fuel = 0
-        else:
-            self.fuel -= distance
-            distance_driven = distance
-        self.odometer += distance_driven
-        return distance_driven
-
-        return distance_driven
-
-mitsy = UnreliableCar("Mitsy", 100, 75)
-print(mitsy.drive(40))
-print(mitsy.__str__())
