+++ Prac08/taxiClass.py
@@ -36,10 +36,10 @@ class Car:
 class Taxi(Car):
     """ specialised version of a Car that includes fare costs """
 
-    def __init__(self, name, fuel, price_per_km):
+    def __init__(self, name, fuel):
         """ initialise a Taxi instance, based on parent class Car """
         super().__init__(name, fuel)
-        self.price_per_km = price_per_km
+        self.price_per_km = 1.2
         self.current_fare_distance = 0
 
     def __str__(self):
